sign-out works, and added total visit count, maybe add todays visit count
