+++ src/components/pomodoro/Pomodoro.js
@@ -1,4 +1,4 @@
-import { Fragment, useState } from "react";
+import { useState } from "react";
 import "react-circular-progressbar/dist/styles.css";
 import PomodoroSetting from "./PomodoroSetting";
 import "react-circular-progressbar/dist/styles.css";
