+++ src/database/data.sql
@@ -1,1 +1,4 @@
 -- This is used when creating a dummy data
+
+INSERT INTO "visits" ("totalvisits", "todayvisits")
+VALUES (0, 0);
